Added publicationDate in Candidate.java
