--- nvi-commons/src/main/java/no/sikt/nva/nvi/common/model/Username.java
@@ -1,5 +1,0 @@
-package no.sikt.nva.nvi.common.model;
-
-public record Username(String value) {
-
-}
