+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/model/business/Institution.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.nvi.common.model;
+package no.sikt.nva.nvi.common.model.business;
 
 import java.net.URI;
